requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0
pathlib==1.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
